requirements.txt
# Core Flask and Server
flask
gunicorn

# Database and ORM
psycopg2-binary
Flask-SQLAlchemy

# Authentication, Serialization, and Utilities
python-jose[cryptography]
Flask-Marshmallow
marshmallow-sqlalchemy
Flask-Cors